requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.14
pyautogui==0.9.54
pyyaml==6.0
psutil==5.9.5
requests==2.31.0